Refactor water management UI by restructuring modal components, introducing a new time picker, and enhancing duration input handling. Remove obsolete changeTimeModal and improve code organization for better maintainability.
