+++ lib/widgets/water_widget/change_time_modal/button_confirm.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:hexcolor/hexcolor.dart';
+
+class ButtonConfirm extends StatelessWidget {
+  const ButtonConfirm({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Expanded(
+          child: Center(
+            child: InkWell(
+              onTap: () {
+                // Agar then() di bawah bisa membedakan antara
+                // dismiss dan confirm.
+                Navigator.of(context).pop(true);
+              },
+              child: Container(
+                height: 40.h,
+                alignment: Alignment.center,
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(8),
+                  color: HexColor('#45D695'),
+                ),
+                child: Text(
+                  'Confirm',
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 14.sp,
+                    fontWeight: FontWeight.w600,
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
