+++ lib/widgets/water_widget/change_time_modal/content_modal/content_change_time_modal.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:iot_plant_control/controller/water_controller.dart';
+import 'package:iot_plant_control/widgets/water_widget/change_time_modal/content_modal/duration_change.dart';
+import 'package:iot_plant_control/widgets/water_widget/change_time_modal/content_modal/time_picker_change.dart';
+
+class ContentChangeTimeModal extends StatelessWidget {
+  const ContentChangeTimeModal({
+    super.key,
+    required this.waterController,
+  });
+
+  final WaterController waterController;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        TimePickerChange(waterController: waterController),
+        DurationChange(waterController: waterController),
+      ],
+    );
+  }
+}
